refactor: move all non-compound components to components folder
